+++ Reminder/HotbarLockReminder.cs
@@ -24,7 +24,7 @@ namespace RemindMe.Reminder {
             return Marshal.ReadByte(actionBar, 0x23F) == 0;
         }
 
-        public override ushort GetIconID(DalamudPluginInterface pluginInterface, RemindMe plugin, MonitorDisplay display) {
+        public override uint GetIconID(DalamudPluginInterface pluginInterface, RemindMe plugin, MonitorDisplay display) {
             return 5; //60840;
         }
